Changed storage of  amplitude and phase into 16 bit values from 32

+++ qt_quick_app/sweeper.cpp
@@ -84,8 +84,8 @@ void Sweeper::callback_cmd_frequency_sweep(QByteArray return_bytes) {
     m_phase_data.clear();
     for (int i = Num_Meta_Data_Bytes; i < return_bytes.size(); i += 4) {
         current_frequency = m_current_resonant_frequency - m_boundaries[m_repetitions_counter] + m_step_sizes[m_repetitions_counter] * ((i - 2) / 4);
-        quint32 amplitude_value = bytes_to_word(quint8(return_bytes[i]), quint8(return_bytes[i + 1]));
-        quint32 phase_value = bytes_to_word(quint8(return_bytes[i + 2]), quint8(return_bytes[i + 3]));
+        quint16 amplitude_value = bytes_to_word(quint8(return_bytes[i]), quint8(return_bytes[i + 1]));
+        quint16 phase_value = bytes_to_word(quint8(return_bytes[i + 2]), quint8(return_bytes[i + 3]));
         m_amplitude_data.append(QPointF(current_frequency, double(amplitude_value) * ADC::SCALE_FACTOR));
         m_phase_data.append(QPointF(current_frequency, double(phase_value) * ADC::SCALE_FACTOR));
 //        qDebug() << QString().sprintf("%2p", quint8(return_bytes[i])) << QString().sprintf("%2p", quint8(return_bytes[i + 1])) << double(amplitude_value * ADC::SCALE_FACTOR) << QString().sprintf("%2p", quint8(return_bytes[i + 2])) << QString().sprintf("%2p", quint8(return_bytes[i + 3])) << double(phase_value * ADC::SCALE_FACTOR);
